backend/requirements.txt
# Django Core
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.4,<6.0
django-cors-headers>=4.6,<5.0

# Database
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.3,<3.0

# Caching & Task Queue
django-redis>=5.4,<6.0
celery>=5.4,<6.0
redis>=5.2,<6.0

# Payments
stripe>=11.0,<12.0

# Storage
django-storages[s3]>=1.14,<2.0
boto3>=1.35,<2.0

# API Documentation
drf-spectacular>=0.28,<1.0

# Security
django-ratelimit>=4.1,<5.0
argon2-cffi>=23.1,<24.0

# Environment
python-decouple>=3.8,<4.0

# WSGI Server
gunicorn>=23.0,<24.0

# Email
postmarker>=1.0,<2.0

# Monitoring
sentry-sdk[django]>=2.19,<3.0
python-json-logger>=3.0,<4.0

# Image Processing
Pillow>=11.0,<12.0

# Development
django-extensions>=3.2,<4.0
django-silk>=5.3,<6.0

# Testing
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=6.0,<7.0
factory-boy>=3.3,<4.0
